Only return named values
It appears that a bug in Zabbix [1] could be limiting the output to 64KB, which
can be a problem for large core systems. As a bit of a debug, attempt to reduce
the amount of output by only including the named values, which is primarily
what I am interested in.

[1]: https://support.zabbix.com/browse/ZBX-5863