Rework the display code for undefined values
I ran into a problem with ERR and MIS with our Zabbix in that it had so many
values that were unsupported that it gave me the typical "more than 100 items
missing" alert. As a work around, set non-existant values to 0. Also, having a
zero is good for those looking to use a trigger as you wouldn't want to use the
CPU0 value across the board as it could trigger hundreds of times for a multi
CPU, high core system.